+++ modules/apps/portal/portal-upload-impl/src/main/java/com/liferay/portal/upload/internal/UploadServletRequestImpl.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portal.upload;
+package com.liferay.portal.upload.internal;
 
 import com.liferay.petra.string.StringBundler;
 import com.liferay.portal.kernel.io.ByteArrayFileInputStream;
@@ -20,6 +20,8 @@ import com.liferay.portal.kernel.util.ProgressTracker;
 import com.liferay.portal.kernel.util.ServiceProxyFactory;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.util.WebKeys;
+import com.liferay.portal.upload.LiferayServletRequest;
+import com.liferay.portal.upload.ServletFileUpload;
 
 import java.io.File;
 import java.io.IOException;
